+++ lib/realize/format/string_template.rb
@@ -1,5 +1,12 @@
 # frozen_string_literal: true
 
+#
+# Copyright (c) 2020-present, Blue Marble Payroll, LLC
+#
+# This source code is licensed under the MIT license found in the
+# LICENSE file in the root directory of this source tree.
+#
+
 module Realize
   class Format
     # Use an expression as a template and string interpolate it using the Stringento library.