+++ spec/realize/format/substring_spec.rb
@@ -1,5 +1,12 @@
 # frozen_string_literal: true
 
+#
+# Copyright (c) 2020-present, Blue Marble Payroll, LLC
+#
+# This source code is licensed under the MIT license found in the
+# LICENSE file in the root directory of this source tree.
+#
+
 require 'spec_helper'
 
 RSpec.describe Realize::Format::Substring do